AI/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioresponses==0.7.6
aiosignal==1.3.1
attrs==24.2.0
backoff==2.2.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.5.0
humanfriendly==10.0
idna==3.10
imageio==2.36.1
inference-sdk==0.9.17
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
onnxruntime==1.20.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
propcache==0.2.0
protobuf==5.29.0
py-cpuinfo==9.0.0
PyMatting==1.1.13
pyparsing==3.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rembg==2.0.60
requests==2.32.3
rpds-py==0.21.0
scikit-image==0.24.0
scipy==1.14.1
six==1.16.0
supervision==0.24.0
sympy==1.13.3
tifffile==2024.9.20
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.2
yarl==1.17.1
